requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
bootstrap4==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-six==1.0.4
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.3
pygame==1.9.6
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==4.1.3
wrapt==1.11.2
xlwt==1.3.0
django-storages==1.7.2
boto3==1.10.14 
botocore==1.13.14 
docutils==0.15.2 
jmespath==0.9.4 
python-dateutil==2.8.0 
s3transfer==0.2.1beautifulsoup4==4.8.1
bootstrap4==0.1.0
boto3==1.10.14
botocore==1.13.14
certifi==2019.9.11
chardet==3.0.4
clang==6.0.0.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-baker==0.22
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-s3direct==1.1.5
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.3
pygame==1.9.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.3
boto==2.49.0
